Support Google Analytics 4 using hugo internal template (#76)
* Support Google Analytics 4 using hugo internal template

Use https://gohugo.io/templates/internal/#google-analytics
in footer partial. Update sample-config.toml; googleAnalytics
now needs to be top level rather than under [params] for the
internal template to work. This might be a concern for folks
upgrading their Hugo-Octopress theme.

Fixes: #75

* remove duplicate google analytics tag from header

It's already in the footer, and
https://gohugo.io/templates/internal/#use-the-google-analytics-template
states that "The async template is not suitable for Google Analytics 4."
which is the non-depracated version of GA.

Keep the sync script in the footer to not block page loading.